+++ Command/MigrateCommand.php
@@ -73,6 +73,8 @@ EOT
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $start = microtime(true);
+
         $this->setOutput($output);
         $this->setVerbosity($output->getVerbosity());
 
@@ -121,12 +123,17 @@ EOT
             }
         }
 
+        $executed = 0;
+        $failed = 0;
+        $skipped = 0;
+
         /** @var MigrationDefinition $migrationDefinition */
         foreach ($toExecute as $name => $migrationDefinition) {
 
             // let's skip migrations that we know are invalid - user was warned and he decided to proceed anyway
             if ($migrationDefinition->status == MigrationDefinition::STATUS_INVALID) {
                 $output->writeln("<comment>Skipping $name</comment>\n");
+                $skipped++;
                 continue;
             }
 
@@ -176,9 +183,12 @@ EOT
                         throw new \Exception($errorMsg);
                     }
 
+                    $executed++;
+
                 } catch (\Exception $e) {
                     if ($input->getOption('ignore-failures')) {
                         $output->writeln("\n<error>Migration failed! Reason: " . $e->getMessage() . "</error>\n");
+                        $failed++;
                         continue;
                     }
                     $output->writeln("\n<error>Migration aborted! Reason: " . $e->getMessage() . "</error>");
@@ -193,9 +203,11 @@ EOT
                         !$input->getOption('no-transactions'),
                         $input->getOption('default-language')
                     );
+                    $executed++;
                 } catch (\Exception $e) {
                     if ($input->getOption('ignore-failures')) {
                         $output->writeln("\n<error>Migration failed! Reason: " . $e->getMessage() . "</error>\n");
+                        $failed++;
                         continue;
                     }
                     $output->writeln("\n<error>Migration aborted! Reason: " . $e->getMessage() . "</error>");
@@ -203,14 +215,21 @@ EOT
                 }
 
             }
-
-            $this->writeln('');
         }
 
         if ($input->getOption('clear-cache')) {
             $command = $this->getApplication()->find('cache:clear');
             $inputArray = new ArrayInput(array('command' => 'cache:clear'));
             $command->run($inputArray, $output);
+        }
+
+        $time = microtime(true) - $start;
+        $this->writeln("Executed $executed migrations, failed $failed, skipped $skipped");
+        if ($input->getOption('separate-process')) {
+            // in case of using subprocesses, we can not measure max memory used
+            $this->writeln("Time taken: ".sprintf('%.2f', $time)." secs");
+        } else {
+            $this->writeln("Time taken: ".sprintf('%.2f', $time)." secs, memory: ".sprintf('%.2f', (memory_get_peak_usage(true) / 1000000)). ' MB');
         }
     }